backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
apscheduler
pytest
pytest-asyncio